+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased](https://github.com/swark-io/swark/compare/v1.3.0...main)
+## [Unreleased](https://github.com/swark-io/swark/compare/v1.3.3...main)
 
 ### Added
 
@@ -18,6 +18,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 -   N/A
+
+## [1.3.3](https://github.com/swark-io/swark/compare/v1.3.0...v1.3.3) - 2024-12-31
+
+### Changed
+
+-   Show progress in VS Code when prompting the LLM.
 
 ## [1.3.0](https://github.com/swark-io/swark/compare/v1.2.0...v1.3.0) - 2024-12-26